requirements.txt
Django==1.10.5
Pillow
Bottlenose
lxml
python-dateutil
python-amazon-simple-product-api
requests